# Use hmset and hgetall to pass args