--- a/.travis.yml
+++ b/.travis.yml
@@ -11,7 +11,7 @@ services:
   - redis-server
 script:
   - sudo ./lua/install.sh 
-  - ./lua51/bin/busted lua/test/*
+  - ./tmp/lua51/bin/busted lua/test/*
   - bundle exec rake rubocop
   - COVERBAND_HASH_REDIS_STORE=t bundle exec rake
   - COVERBAND_HASH_REDIS_STORE=t bundle exec rake forked_tests