--- a/.travis.yml
+++ b/.travis.yml
@@ -10,6 +10,8 @@ gemfile:
 services:
   - redis-server
 script:
+  - sudo ./lua/install.sh 
+  - ./lua51/bin/busted lua/test/*
   - bundle exec rake rubocop
   - COVERBAND_HASH_REDIS_STORE=t bundle exec rake
   - COVERBAND_HASH_REDIS_STORE=t bundle exec rake forked_tests