--- a/lua/install.sh
+++ b/lua/install.sh
@@ -1,8 +1,9 @@
 #!/bin/bash
 
+LUA_DIR="./tmp/lua51"
 pip install hererocks
-hererocks lua51 -l5.1 -rlatest
-source lua51/bin/activate
+hererocks $LUA_DIR -l5.1 -rlatest
+source $LUA_DIR/bin/activate
 lua -v
 for i in luacov busted redis-lua inspect; do 
   luarocks install $i;