--- a/lua/lib/persist-coverage.lua
+++ b/lua/lib/persist-coverage.lua
@@ -1,17 +1,25 @@
-local first_updated_at = table.remove(ARGV, 1)
-local last_updated_at = table.remove(ARGV, 1)
-local file = table.remove(ARGV, 1)
-local file_hash = table.remove(ARGV, 1)
-local ttl = table.remove(ARGV, 1)
-local file_length = table.remove(ARGV, 1)
-local hash_key = table.remove(KEYS, 1)
+local hash_values = redis.call('HGETALL', KEYS[1])
+function remove(key)
+  local val = hash_values[key]
+  hash_values[key] = nil
+  return val
+end
+
+redis.call('DEL', KEYS[1])
+local first_updated_at = remove('first_updated_at')
+local last_updated_at = remove('last_updated_at')
+local file = remove('file')
+local file_hash = remove('file_hash')
+local ttl = remove('ttl')
+local file_length = remove('file_length')
+local hash_key = remove('hash_key')
 redis.call('HMSET', hash_key, 'last_updated_at', last_updated_at, 'file', file, 'file_hash', file_hash, 'file_length', file_length)
 redis.call('HSETNX', hash_key, 'first_updated_at', first_updated_at)
-for i, key in ipairs(KEYS) do
-  if ARGV[i] == '-1' then
-    redis.call("HSET", hash_key, key, ARGV[i])
+for line, coverage in pairs(hash_values) do
+  if coverage  == '-1' then
+    redis.call("HSET", hash_key, line, coverage)
   else
-    redis.call("HINCRBY", hash_key, key, ARGV[i])
+    redis.call("HINCRBY", hash_key, line, coverage)
   end
 end
 if ttl ~= '-1' then