--- a/lua/test/bootstrap.lua
+++ b/lua/test/bootstrap.lua
@@ -1,0 +1,4 @@
+inspect = require "inspect"
+require './lua/test/redis-call'
+
+