--- a/lua/test/harness.lua
+++ b/lua/test/harness.lua
@@ -1,5 +1,4 @@
-inspect = require "inspect"
-require './lua/test/redis-call'
+require './lua/test/bootstrap'
 
 
 KEYS = {}