--- a/lua/test/test-persist-coverage.lua
+++ b/lua/test/test-persist-coverage.lua
@@ -17,9 +17,25 @@ describe("incr-and-stor", function()
   it("should add single items", function()
     local first_updated_at = "1569453853"
     local last_updated_at = first_updated_at
-    local args = { first_updated_at, last_updated_at, "./dog.rb", "abcd", '-1', 3, 0, 1, 2 }
-    local keys = {"coverband_hash_3_3.coverband_test.runtime../dog.rb.abcd", 0, 1, 2 }
-    call_redis_script('persist-coverage.lua',  keys ,  args );
+
+    local key = 'hash_key'
+    redis.call( 'hmset', key, 
+    'hash_key', "coverband_hash_3_3.coverband_test.runtime../dog.rb.abcd",
+    'first_updated_at', 
+    first_updated_at, 
+    'last_updated_at', 
+    last_updated_at, 
+    'file', "./dog.rb", 
+    'file_hash', 'abcd', 
+    'ttl', '-1', 
+    'file_length', 3, 
+    0, 0, 
+    1, 1, 
+    2, 2)
+
+
+
+    call_redis_script('persist-coverage.lua',  { key },  {});
     local results = redis.call('HGETALL', "coverband_hash_3_3.coverband_test.runtime../dog.rb.abcd")
     assert.are.same({
       ["0"] = "0",
@@ -31,11 +47,23 @@ describe("incr-and-stor", function()
       first_updated_at = first_updated_at ,
       last_updated_at = last_updated_at
     }, results)
+    
+    assert.is_false(false, redis.call('exists', key))
 
     last_updated_at = "1569453953"
-    args = { last_updated_at, last_updated_at, "./dog.rb", "abcd", '-1', 3, 1, 1, 1 }
-    keys = {"coverband_hash_3_3.coverband_test.runtime../dog.rb.abcd", 0, 1, 2 }
-    call_redis_script('persist-coverage.lua',  keys ,  args );
+    redis.call( 'hmset', key, 
+    'hash_key', "coverband_hash_3_3.coverband_test.runtime../dog.rb.abcd",
+    'first_updated_at', first_updated_at, 
+    'last_updated_at', last_updated_at, 
+    'file', "./dog.rb", 
+    'file_hash', 'abcd', 
+    'ttl', '-1', 
+    'file_length', 3, 
+    0, 1, 
+    1, 1, 
+    2, 1)
+
+    call_redis_script('persist-coverage.lua',  { key },  {} );
     results = redis.call('HGETALL', "coverband_hash_3_3.coverband_test.runtime../dog.rb.abcd")
     assert.are.same({
       ["0"] = "1",
@@ -48,5 +76,6 @@ describe("incr-and-stor", function()
       last_updated_at = last_updated_at 
     }, results)
 
+    assert.is_false(false, redis.call('exists', key))
   end)
 end)