Uses about 25% less cpu on redis instance:

Before on master:

![image](https://user-images.githubusercontent.com/96786/66274871-f95a8e80-e850-11e9-83c6-4009ae542293.png)

```
  coverband git:(master) COVERBAND_HASH_REDIS_STORE=t bundle exec rake benchmarks:redis_reporting
Warming up --------------------------------------
   store_reports_all     1.000  i/100ms
Calculating -------------------------------------
   store_reports_all      3.871  (±25.8%) i/s -     56.000  in  15.148240s
Warming up --------------------------------------
store_reports_subset     7.000  i/100ms
Calculating -------------------------------------
store_reports_subset    109.200  (±12.8%) i/s -      2.149k in  20.008229s
```

On this branch:

![image](https://user-images.githubusercontent.com/96786/66274898-55251780-e851-11e9-9de4-8a17aaa90ec2.png)

```
➜  coverband git:(change_to_hash) COVERBAND_HASH_REDIS_STORE=t bundle exec rake benchmarks:redis_reporting
runs benchmarks on reporting large sets of files to redis
Warming up --------------------------------------
   store_reports_all     1.000  i/100ms
Calculating -------------------------------------
   store_reports_all      2.319  (± 0.0%) i/s -     35.000  in  15.179692s
Warming up --------------------------------------
store_reports_subset     6.000  i/100ms
Calculating -------------------------------------
store_reports_subset     60.083  (±10.0%) i/s -      1.194k in  20.056954s
```

Seems a little slower